requirements.txt
rpl==1.15.5
ruff==0.0.292
nbqa==1.7.0
texttable==1.6.7
jupyter==1.0.0
pytest==7.4.2
pydantic==2.5.2
PyYAML==6.0.1

# Numerics
numpy==1.26.2
scipy==1.11.4

# Visualization
matplotlib==3.8.2

# Image Processing
imageio==2.33.0
opencv-python==4.8.1.78
Pillow==10.1.0
scikit-image==0.22.0

# Machine Learning
pandas==2.1.3
scikit-learn==1.3.2
torch==2.1.1
torchvision==0.16.1
torchinfo==1.8.0
torchviz==0.0.2
albumentations==1.3.1
torchmetrics==1.2.0
lightning==2.1.2
tensorrt==8.6.1.post1
tensorboard==2.15.1
torch-tb-profiler==0.4.3
#transformers==4.35.2  # Beware: Adding this causes pytest deprecation warnings.
cityscapesScripts==2.2.2

# Other
watermark==2.4.3
